Add support for HTTPS
It was not possible to interrogate OpenVeo Web Service in HTTPS. Also
url of the Web Service was divided into host and port, it now expects
the full Web Service url. Finally the full Web Service url was used with
get / post / put / delete methods to make a call to the Web Service, it
now expects only the end point with query parameters.